+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/HlsChunkSource.java
@@ -35,6 +35,7 @@ import androidx.media3.common.util.Util;
 import androidx.media3.datasource.DataSource;
 import androidx.media3.datasource.DataSpec;
 import androidx.media3.datasource.TransferListener;
+import androidx.media3.exoplayer.LoadingInfo;
 import androidx.media3.exoplayer.SeekParameters;
 import androidx.media3.exoplayer.analytics.PlayerId;
 import androidx.media3.exoplayer.hls.playlist.HlsMediaPlaylist;
@@ -364,10 +365,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
    * but the end of the stream has not been reached, {@link HlsChunkHolder#playlistUrl} is set to
    * contain the {@link Uri} that refers to the playlist that needs refreshing.
    *
-   * @param playbackPositionUs The current playback position relative to the period start in
-   *     microseconds. If playback of the period to which this chunk source belongs has not yet
-   *     started, the value will be the starting position in the period minus the duration of any
-   *     media in previous periods still to be played.
+   * @param loadingInfo The {@link LoadingInfo} when loading request is made.
    * @param loadPositionUs The current load position relative to the period start in microseconds.
    * @param queue The queue of buffered {@link HlsMediaChunk}s.
    * @param allowEndOfStream Whether {@link HlsChunkHolder#endOfStream} is allowed to be set for
@@ -376,13 +374,14 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
    * @param out A holder to populate.
    */
   public void getNextChunk(
-      long playbackPositionUs,
+      LoadingInfo loadingInfo,
       long loadPositionUs,
       List<HlsMediaChunk> queue,
       boolean allowEndOfStream,
       HlsChunkHolder out) {
     @Nullable HlsMediaChunk previous = queue.isEmpty() ? null : Iterables.getLast(queue);
     int oldTrackIndex = previous == null ? C.INDEX_UNSET : trackGroup.indexOf(previous.trackFormat);
+    long playbackPositionUs = loadingInfo.playbackPositionUs;
     long bufferedDurationUs = loadPositionUs - playbackPositionUs;
     long timeToLiveEdgeUs = resolveTimeToLiveEdgeUs(playbackPositionUs);
     if (previous != null && !independentSegments) {