+++ libraries/exoplayer_hls/src/main/java/androidx/media3/exoplayer/hls/HlsSampleStreamWrapper.java
@@ -761,7 +761,7 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
     }
     nextChunkHolder.clear();
     chunkSource.getNextChunk(
-        loadingInfo.playbackPositionUs,
+        loadingInfo,
         loadPositionUs,
         chunkQueue,
         /* allowEndOfStream= */ prepared || !chunkQueue.isEmpty(),